requirements.txt
pypdf>=4.0.0
reportlab>=4.0.0
pypdfium2>=4.0.0
Pillow>=9.0.0
PyMuPDF>=1.23.0
pytest>=7.0.0
pytest-cov>=4.0.0
